@@ -27,10 +27,10 @@
     "@babel/preset-typescript": "^7.3.3",
     "@types/classnames": "^2.2.9",
     "@types/history": "^4.7.2",
-    "@types/react": "^16.8.17",
+    "@types/react": "^17.0.0",
     "@types/react-burger-menu": "^2.2.5",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-router-dom": "^5.1.3",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.1.6",
     "@types/react-scroll": "^1.5.4",
     "@types/react-select": "^3.0.5",
     "@types/reactstrap": "^8.0.4",
@@ -80,14 +80,14 @@
     "classnames": "^2.2.6",
     "express": "^4.17.1",
     "mixpanel-browser": "^2.29.1",
-    "react": "^16.8.6",
+    "react": "^17.0.1",
     "react-burger-menu": "^2.6.11",
     "react-device-detect": "^1.9.10",
-    "react-dom": "^16.8.6",
+    "react-dom": "^17.0.1",
     "react-icons": "^3.7.0",
     "react-particles-js": "^2.7.0",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.0.0",
+    "react-router-dom": "^5.2.0",
     "react-scroll": "^1.7.14",
     "react-scroll-parallax": "^2.2.0",
     "react-select": "^3.0.8",
